requirements.txt
pandas>=2.3.3
ipykernel>=7.0.1
pyyaml>=6.0.3